+++ src/js/control-bar/progress-control/seek-handle.js
@@ -5,7 +5,7 @@ import * as Lib from '../../lib.js';
  * The Seek Handle shows the current position of the playhead during playback,
  * and can be dragged to adjust the playhead.
  *
- * @param {vjs.Player|Object} player
+ * @param {Player|Object} player
  * @param {Object=} options
  * @constructor
  */
@@ -26,7 +26,7 @@ class SeekHandle extends SliderHandle {
 
   updateContent() {
     let time = (this.player_.scrubbing) ? this.player_.getCache().currentTime : this.player_.currentTime();
-    this.el_.innerHTML = '<span class="vjs-control-text">' + Lib.formatTime(time, this.player_.duration()) + '</span>';
+    this.el_.innerHTML = `<span class="vjs-control-text">${Lib.formatTime(time, this.player_.duration())}</span>`;
   }
 
 }