+++ src/js/control-bar/time-controls/time-divider.js
@@ -1,11 +1,11 @@
-import Component from '../component.js';
+import Component from '../../component.js';
 
 /**
  * The separator between the current time and duration
  *
  * Can be hidden if it's not needed in the design.
  *
- * @param {vjs.Player|Object} player
+ * @param {Player|Object} player
  * @param {Object=} options
  * @constructor
  */