+++ src/js/options.js
@@ -3,8 +3,8 @@ import window from 'global/window';
 let navigator = window.navigator;
 
 /**
- * Global Player instance options, surfaced from vjs.Player.prototype.options_
- * vjs.options = vjs.Player.prototype.options_
+ * Global Player instance options, surfaced from Player.prototype.options_
+ * options = Player.prototype.options_
  * All options should use string keys so they avoid
  * renaming by closure compiler
  * @type {Object}