+++ src/js/slider/slider-handle.js
@@ -4,7 +4,7 @@ import * as Lib from '../lib.js';
 /**
  * SeekBar Behavior includes play progress bar, and seek handle
  * Needed so it can determine seek position based on handle position/size
- * @param {vjs.Player|Object} player
+ * @param {Player|Object} player
  * @param {Object=} options
  * @constructor
  */
@@ -16,7 +16,7 @@ class SliderHandle extends Component {
     // Add the slider element class to all sub classes
     props.className = props.className + ' vjs-slider-handle';
     props = Lib.obj.merge({
-      innerHTML: '<span class="vjs-control-text">'+(this.defaultValue || 0)+'</span>'
+      innerHTML: `<span class="vjs-control-text">${this.defaultValue || 0}</span>`
     }, props);
 
     return super.createEl('div', props);