+++ src/js/tracks/text-track-settings.js
@@ -168,118 +168,120 @@ function setSelectedOption(target, value) {
 }
 
 function captionOptionsMenuTemplate() {
-  return '<div class="vjs-tracksettings">' +
-      '<div class="vjs-tracksettings-colors">' +
-        '<div class="vjs-fg-color vjs-tracksetting">' +
-            '<label class="vjs-label">Foreground</label>' +
-            '<select>' +
-              '<option value="">---</option>' +
-              '<option value="#FFF">White</option>' +
-              '<option value="#000">Black</option>' +
-              '<option value="#F00">Red</option>' +
-              '<option value="#0F0">Green</option>' +
-              '<option value="#00F">Blue</option>' +
-              '<option value="#FF0">Yellow</option>' +
-              '<option value="#F0F">Magenta</option>' +
-              '<option value="#0FF">Cyan</option>' +
-            '</select>' +
-            '<span class="vjs-text-opacity vjs-opacity">' +
-              '<select>' +
-                '<option value="">---</option>' +
-                '<option value="1">Opaque</option>' +
-                '<option value="0.5">Semi-Opaque</option>' +
-              '</select>' +
-            '</span>' +
-        '</div>' + // vjs-fg-color
-        '<div class="vjs-bg-color vjs-tracksetting">' +
-            '<label class="vjs-label">Background</label>' +
-            '<select>' +
-              '<option value="">---</option>' +
-              '<option value="#FFF">White</option>' +
-              '<option value="#000">Black</option>' +
-              '<option value="#F00">Red</option>' +
-              '<option value="#0F0">Green</option>' +
-              '<option value="#00F">Blue</option>' +
-              '<option value="#FF0">Yellow</option>' +
-              '<option value="#F0F">Magenta</option>' +
-              '<option value="#0FF">Cyan</option>' +
-            '</select>' +
-            '<span class="vjs-bg-opacity vjs-opacity">' +
-                '<select>' +
-                  '<option value="">---</option>' +
-                  '<option value="1">Opaque</option>' +
-                  '<option value="0.5">Semi-Transparent</option>' +
-                  '<option value="0">Transparent</option>' +
-                '</select>' +
-            '</span>' +
-        '</div>' + // vjs-bg-color
-        '<div class="window-color vjs-tracksetting">' +
-            '<label class="vjs-label">Window</label>' +
-            '<select>' +
-              '<option value="">---</option>' +
-              '<option value="#FFF">White</option>' +
-              '<option value="#000">Black</option>' +
-              '<option value="#F00">Red</option>' +
-              '<option value="#0F0">Green</option>' +
-              '<option value="#00F">Blue</option>' +
-              '<option value="#FF0">Yellow</option>' +
-              '<option value="#F0F">Magenta</option>' +
-              '<option value="#0FF">Cyan</option>' +
-            '</select>' +
-            '<span class="vjs-window-opacity vjs-opacity">' +
-                '<select>' +
-                  '<option value="">---</option>' +
-                  '<option value="1">Opaque</option>' +
-                  '<option value="0.5">Semi-Transparent</option>' +
-                  '<option value="0">Transparent</option>' +
-                '</select>' +
-            '</span>' +
-        '</div>' + // vjs-window-color
-      '</div>' + // vjs-tracksettings
-      '<div class="vjs-tracksettings-font">' +
-        '<div class="vjs-font-percent vjs-tracksetting">' +
-          '<label class="vjs-label">Font Size</label>' +
-          '<select>' +
-            '<option value="0.50">50%</option>' +
-            '<option value="0.75">75%</option>' +
-            '<option value="1.00" selected>100%</option>' +
-            '<option value="1.25">125%</option>' +
-            '<option value="1.50">150%</option>' +
-            '<option value="1.75">175%</option>' +
-            '<option value="2.00">200%</option>' +
-            '<option value="3.00">300%</option>' +
-            '<option value="4.00">400%</option>' +
-          '</select>' +
-        '</div>' + // vjs-font-percent
-        '<div class="vjs-edge-style vjs-tracksetting">' +
-          '<label class="vjs-label">Text Edge Style</label>' +
-          '<select>' +
-            '<option value="none">None</option>' +
-            '<option value="raised">Raised</option>' +
-            '<option value="depressed">Depressed</option>' +
-            '<option value="uniform">Uniform</option>' +
-            '<option value="dropshadow">Dropshadow</option>' +
-          '</select>' +
-        '</div>' + // vjs-edge-style
-        '<div class="vjs-font-family vjs-tracksetting">' +
-          '<label class="vjs-label">Font Family</label>' +
-          '<select>' +
-            '<option value="">Default</option>' +
-            '<option value="monospaceSerif">Monospace Serif</option>' +
-            '<option value="proportionalSerif">Proportional Serif</option>' +
-            '<option value="monospaceSansSerif">Monospace Sans-Serif</option>' +
-            '<option value="proportionalSansSerif">Proportional Sans-Serif</option>' +
-            '<option value="casual">Casual</option>' +
-            '<option value="script">Script</option>' +
-            '<option value="small-caps">Small Caps</option>' +
-          '</select>' +
-        '</div>' + // vjs-font-family
-      '</div>' +
-    '</div>' +
-    '<div class="vjs-tracksettings-controls">' +
-      '<button class="vjs-default-button">Defaults</button>' +
-      '<button class="vjs-done-button">Done</button>' +
-    '</div>';
+  let template = `<div class="vjs-tracksettings">
+      <div class="vjs-tracksettings-colors">
+        <div class="vjs-fg-color vjs-tracksetting">
+            <label class="vjs-label">Foreground</label>
+            <select>
+              <option value="">---</option>
+              <option value="#FFF">White</option>
+              <option value="#000">Black</option>
+              <option value="#F00">Red</option>
+              <option value="#0F0">Green</option>
+              <option value="#00F">Blue</option>
+              <option value="#FF0">Yellow</option>
+              <option value="#F0F">Magenta</option>
+              <option value="#0FF">Cyan</option>
+            </select>
+            <span class="vjs-text-opacity vjs-opacity">
+              <select>
+                <option value="">---</option>
+                <option value="1">Opaque</option>
+                <option value="0.5">Semi-Opaque</option>
+              </select>
+            </span>
+        </div> <!-- vjs-fg-color -->
+        <div class="vjs-bg-color vjs-tracksetting">
+            <label class="vjs-label">Background</label>
+            <select>
+              <option value="">---</option>
+              <option value="#FFF">White</option>
+              <option value="#000">Black</option>
+              <option value="#F00">Red</option>
+              <option value="#0F0">Green</option>
+              <option value="#00F">Blue</option>
+              <option value="#FF0">Yellow</option>
+              <option value="#F0F">Magenta</option>
+              <option value="#0FF">Cyan</option>
+            </select>
+            <span class="vjs-bg-opacity vjs-opacity">
+                <select>
+                  <option value="">---</option>
+                  <option value="1">Opaque</option>
+                  <option value="0.5">Semi-Transparent</option>
+                  <option value="0">Transparent</option>
+                </select>
+            </span>
+        </div> <!-- vjs-bg-color -->
+        <div class="window-color vjs-tracksetting">
+            <label class="vjs-label">Window</label>
+            <select>
+              <option value="">---</option>
+              <option value="#FFF">White</option>
+              <option value="#000">Black</option>
+              <option value="#F00">Red</option>
+              <option value="#0F0">Green</option>
+              <option value="#00F">Blue</option>
+              <option value="#FF0">Yellow</option>
+              <option value="#F0F">Magenta</option>
+              <option value="#0FF">Cyan</option>
+            </select>
+            <span class="vjs-window-opacity vjs-opacity">
+                <select>
+                  <option value="">---</option>
+                  <option value="1">Opaque</option>
+                  <option value="0.5">Semi-Transparent</option>
+                  <option value="0">Transparent</option>
+                </select>
+            </span>
+        </div> <!-- vjs-window-color -->
+      </div> <!-- vjs-tracksettings -->
+      <div class="vjs-tracksettings-font">
+        <div class="vjs-font-percent vjs-tracksetting">
+          <label class="vjs-label">Font Size</label>
+          <select>
+            <option value="0.50">50%</option>
+            <option value="0.75">75%</option>
+            <option value="1.00" selected>100%</option>
+            <option value="1.25">125%</option>
+            <option value="1.50">150%</option>
+            <option value="1.75">175%</option>
+            <option value="2.00">200%</option>
+            <option value="3.00">300%</option>
+            <option value="4.00">400%</option>
+          </select>
+        </div> <!-- vjs-font-percent -->
+        <div class="vjs-edge-style vjs-tracksetting">
+          <label class="vjs-label">Text Edge Style</label>
+          <select>
+            <option value="none">None</option>
+            <option value="raised">Raised</option>
+            <option value="depressed">Depressed</option>
+            <option value="uniform">Uniform</option>
+            <option value="dropshadow">Dropshadow</option>
+          </select>
+        </div> <!-- vjs-edge-style -->
+        <div class="vjs-font-family vjs-tracksetting">
+          <label class="vjs-label">Font Family</label>
+          <select>
+            <option value="">Default</option>
+            <option value="monospaceSerif">Monospace Serif</option>
+            <option value="proportionalSerif">Proportional Serif</option>
+            <option value="monospaceSansSerif">Monospace Sans-Serif</option>
+            <option value="proportionalSansSerif">Proportional Sans-Serif</option>
+            <option value="casual">Casual</option>
+            <option value="script">Script</option>
+            <option value="small-caps">Small Caps</option>
+          </select>
+        </div> <!-- vjs-font-family -->
+      </div>
+    </div>
+    <div class="vjs-tracksettings-controls">
+      <button class="vjs-default-button">Defaults</button>
+      <button class="vjs-done-button">Done</button>
+    </div>`;
+
+    return template;
 }
 
 export default TextTrackSettings;